% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_foo.R
\name{align_phase}
\alias{align_phase}
\title{Align a vector of phases to a target Circadian Time}
\usage{
align_phase(data, CT = 6, ...)
}
\arguments{
\item{data}{Numeric vector of phases in radians.}

\item{CT}{Numeric; target Circadian Time in hours. Defaults to \code{6}.}

\item{...}{Currently unused.}
}
\value{
Numeric vector of aligned phases in radians.
}
\description{
Align a vector of phases to a target Circadian Time
}
\keyword{internal}
