% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_foo.R
\name{assign_cluster_colors}
\alias{assign_cluster_colors}
\title{Assign colors to clusters}
\usage{
assign_cluster_colors(node_data)
}
\arguments{
\item{node_data}{Data frame with a \code{cluster} factor column.}
}
\value{
A data frame with columns \code{cluster} and \code{color}.
\code{"NULL"} clusters are assigned \code{"grey80"}.
}
\description{
Assign colors to clusters
}
