% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClockCyteR.spatial-package.R
\docType{package}
\name{ClockCyteR.spatial-package}
\alias{ClockCyteR.spatial}
\alias{ClockCyteR.spatial-package}
\title{ClockCyteR.spatial: ClockCyteR.spatial implements spatial analysis of circadian data}
\description{
The ClockCyteR.spatial package enables spatiotemporal analysis of timeseries data extracted from timelapse images across multiple channels. It performs multivariate analysis of circadian parameters starting from period, phase, and amplitude from defined ROIs and can calculate network features (optimised for Suprachiasmatic Nucleus organotypic slices).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cabajr.github.io/ClockCyteR.spatial/}
}

}
\author{
\strong{Maintainer}: Marco Ferrari \email{m.ferrari20@imperial.ac.uk} (\href{https://orcid.org/0000-0001-9407-7970}{ORCID})

}
\keyword{internal}
