% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_foo.R
\name{clockcyter_example_path}
\alias{clockcyter_example_path}
\title{Path to the toy dataset folder}
\usage{
clockcyter_example_path()
}
\value{
the path to the toy dataset folder
}
\description{
Path to the toy dataset folder
}
