% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_foo.R
\name{cluster_stats}
\alias{cluster_stats}
\title{Calculate statistics on each cluster}
\usage{
cluster_stats(node_data)
}
\arguments{
\item{node_data}{table containing period analysis in each node}
}
\description{
Calculate statistics on each cluster
}
\keyword{internal}
