% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_foo.R
\name{importROI}
\alias{importROI}
\title{function to import a single ROI}
\usage{
importROI(path, pixel_fct, convert = TRUE)
}
\arguments{
\item{path}{file path of the ROI to import}

\item{pixel_fct}{the pixel size in microns}

\item{convert}{TRUE/FALSE to convert pixels values to microns}
}
\value{
A sf object containing the coordinates defining the ROI
imported
}
\description{
function to import a single ROI
}
