% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_foo.R
\name{list2table}
\alias{list2table}
\title{Bind a list of data frames into a single data frame}
\usage{
list2table(...)
}
\arguments{
\item{...}{Data frames to combine; all must have the same column names and
number of columns.}
}
\value{
A single data frame with all rows from the input data frames.
}
\description{
Bind a list of data frames into a single data frame
}
\keyword{internal}
