% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_foo.R
\name{normalize_phase}
\alias{normalize_phase}
\title{Normalise phases to hours centred around zero}
\usage{
normalize_phase(angles)
}
\arguments{
\item{angles}{Numeric vector or circular object of phase values in radians.}
}
\value{
Numeric vector of phase values in hours, centred around zero in
the range \code{[-12, 12]}.
}
\description{
Normalise phases to hours centred around zero
}
\keyword{internal}
