% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_foo.R
\name{plot_overlap_heatmap}
\alias{plot_overlap_heatmap}
\title{Cluster segregation heatmap}
\usage{
plot_overlap_heatmap(overlap_matrix)
}
\arguments{
\item{overlap_matrix}{Square numeric matrix of pairwise cluster segregation
scores; row and column names are cluster labels.}
}
\value{
A \code{ggplot} object.
}
\description{
Cluster segregation heatmap
}
