% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_foo.R
\name{prepare_file}
\alias{prepare_file}
\title{prepare_file}
\usage{
prepare_file(file_row, params)
}
\arguments{
\item{file_row}{list of files to analyse, output of index_files function}

\item{params}{params object created by the make_params function}
}
\value{
a list containing the file_id, the ROI, the coordinates of each
grid element, the coordinates of each point as sf elements, the path to
the main folder, and the channels to be analysed
}
\description{
prepare_file
}
