% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_foo.R
\name{setup_example}
\alias{setup_example}
\title{Copy the toy dataset to a writable location}
\usage{
setup_example(dest = file.path(tempdir(), "ClockCyteR_example"))
}
\arguments{
\item{dest}{Character; directory where the toy dataset will be copied.
Defaults to a \code{ClockCyteR_example} subdirectory inside
\code{tempdir()}.}
}
\value{
Character; path to the copied dataset folder (invisibly).
}
\description{
Copy the toy dataset to a writable location
}
