% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_foo.R
\name{summarise_cluster_data}
\alias{summarise_cluster_data}
\title{summarise node data}
\usage{
summarise_cluster_data(node_data)
}
\arguments{
\item{node_data}{Data frame of node-level data including columns
\code{cluster}, \code{phase_crn}, \code{period_crn}, \code{amplitude},
\code{period}, \code{clust_coef}, \code{degree}, and \code{strength}.}
}
\description{
summarise node data
}
\keyword{internal}
